// lib/portainer/schema.d.ts
export interface paths {
// ...
    "/endpoints/{id}/docker/v2/browse/ls": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get the contents of a path (endpoint)
         * @description Use this environment(endpoint) to get contents of a path.
         */
        get: {
            parameters: {
                query: {
                    /** @description Path to the file */
                    path: string;
                    /** @description Volume identifier */
                    volumeID: string;
                };
                header?: never;
                path: {
                    /** @description Environment(Endpoint) identifier */
                    id: number;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Success */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            Name: string;
                            Size: number;
                            Dir: boolean;
                            ModTime: number;
                        }[];
                    };
                };
                /** @description Invalid request */
                400: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
                /** @description Server error */
                500: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}